requirements.txt
pytes==7.4.0
playwright==1.36.0